services/rag-gateway/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pgvector>=0.2.4

# Redis
redis>=5.0.1

# HTTP Clients
httpx>=0.26.0

# AI/ML
sentence-transformers>=2.3.1
torch>=2.1.2
transformers>=4.37.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
lxml>=5.1.0
tiktoken>=0.5.2
aiofiles>=23.2.1

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
structlog>=24.1.0
tenacity>=8.2.3
python-dateutil>=2.8.2

# Testing (optional - install separately if needed)
# pytest
# pytest-asyncio
# pytest-cov

# Code Quality (optional - install separately if needed)
# black
# ruff
# mypy
